+++ CHANGELOG.md
@@ -15,6 +15,7 @@ keybinds if they contained either `{` or `}`. You will now need to escape these 
 - `on_exit` config option to execute a command when exiting rmpc, similar to `on_song_change`
 - Added option to directly clear rating in the `Rate` keybind
 - Added `quit_closes_modal`
+- `remove` to the CLI allowing you to remove songs from the queue
 
 ### Changed